Squashed commit of the following:
commit d3a284b094fe8cff6ff45455927d74afd5f8ca26
Author: Szymon <dembkowski232@gmail.com>
Date:   Thu May 29 00:31:35 2025 +0200

    ready to add swagger

    made some fixes in pom.xml and ide settings

commit 40b9a6a9828302b7f8c651ec8efb847203affea5
Author: Szymon <dembkowski232@gmail.com>
Date:   Tue May 27 01:08:59 2025 +0200

    added exceptionhandler

    additional changes in services

commit c1c17b65c687e4ac6e29572333ccb0f3fa031edc
Author: Szymon <dembkowski232@gmail.com>
Date:   Mon May 26 00:24:36 2025 +0200

    Added size to spider gender validation

commit 8b37bcac683ec8293f27e47341400d5bb4a868a1
Author: Szymon <dembkowski232@gmail.com>
Date:   Mon May 26 00:10:28 2025 +0200

    Entity to DTO conversion

    Added DTO logic, made some fixes in controller and services

+++ src/main/java/com.spidershop/Services/SpiderServiceImpl.java
@@ -1,12 +1,17 @@
 package com.spidershop.Services;
 
+import com.spidershop.Dto.SpiderDto;
 import com.spidershop.Entity.Spider;
+import com.spidershop.Entity.SpiderSize;
+import com.spidershop.Exception.SpiderNotFoundException;
 import com.spidershop.Repository.SpiderRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
+@Transactional
 public class SpiderServiceImpl implements SpiderService {
 
     private final SpiderRepository spiderRepository;
@@ -23,35 +28,50 @@ public class SpiderServiceImpl implements SpiderService {
     @Override
     public Spider getSpiderById(Long id) {
         return spiderRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Spider not found"));
+                .orElseThrow(() -> new SpiderNotFoundException("Spider not found with id: " + id));
     }
 
     @Override
-    public Spider createSpider(Spider spider) {
+    public Spider createSpider(SpiderDto spiderDto) {
+        validateSexBasedOnSize(spiderDto);
+        Spider spider = new Spider();
+        spider.setGenus(spiderDto.getGenus());
+        spider.setSpecies(spiderDto.getSpecies());
+        spider.setPrice(spiderDto.getPrice());
+        spider.setQuantity(spiderDto.getQuantity());
+        spider.setSize(spiderDto.getSize());
+        spider.setSex(spiderDto.getSex());
         return spiderRepository.save(spider);
     }
 
     @Override
-    public Spider updateSpider(Long id, Spider updatedSpider) {
+    public Spider updateSpider(Long id, SpiderDto spiderDto) {
+        validateSexBasedOnSize(spiderDto);
+        Spider spider = spiderRepository.findById(id)
+                .orElseThrow(() -> new SpiderNotFoundException("Spider not found with id: " + id));
 
-        Spider spider = spiderRepository.getSpiderById(id)
-                .orElseThrow(() -> new RuntimeException("Spider not found with id: " + id));
-
-        spider.setGenus(updatedSpider.getGenus());
-        spider.setSpecies(updatedSpider.getSpecies());
-        spider.setSize(updatedSpider.getSize());
-        spider.setSex(updatedSpider.getSex());
-        spider.setPrice(updatedSpider.getPrice());
-        spider.setQuantity(updatedSpider.getQuantity());
-        spider.setDescription(updatedSpider.getDescription());
-        spider.setAvailable(updatedSpider.isAvailable());
-
+        spider.setGenus(spiderDto.getGenus());
+        spider.setSpecies(spiderDto.getSpecies());
+        spider.setPrice(spiderDto.getPrice());
+        spider.setQuantity(spiderDto.getQuantity());
+        spider.setSize(spiderDto.getSize());
+        spider.setSex(spiderDto.getSex());
         return spiderRepository.save(spider);
     }
 
     @Override
     public void deleteSpider(Long id) {
+        if (!spiderRepository.existsById(id)) {
+            throw new SpiderNotFoundException("Spider not found with id: " + id);
+        }
         spiderRepository.deleteById(id);
     }
 
+    private void validateSexBasedOnSize(SpiderDto spiderDto){
+        if(spiderDto.getSize().compareTo(SpiderSize.Juvenile) >= 0) {
+            if(spiderDto.getSex() == null){
+                throw new IllegalArgumentException("Sex must be specified for spiders of size Junvenile or greater.");
+            }
+        }
+    }
 }
